:recycle: [#54] Refactor zaak table
