+++ src/archiefvernietigingscomponent/js/components/constants.js
@@ -1,0 +1,6 @@
+const ARCHIEFNOMINATIE_CHOICES = [
+    ["blijvend_bewaren", "blijvend bewaren"],
+    ["vernietigen", "vernietigen"]
+];
+
+export {ARCHIEFNOMINATIE_CHOICES};
