+++ src/archiefvernietigingscomponent/js/components/destruction-list-create/zaken-table.js
@@ -2,18 +2,10 @@ import React, { useState } from 'react';
 import { CheckboxInput } from "../../forms/inputs";
 import { Loader } from '../loader';
 import ErrorMessage from '../ErrorMessage';
+import ZaakRecord from "../no-archive-date/ZaakRecord";
 
 
-const displayZaaktype = (zaaktype) => {
-    return (
-        <span className="zaak-record__zaaktype" title={`versie ${zaaktype.versiedatum}`}>
-            {zaaktype.omschrijving}
-        </span>
-    );
-}
-
-
-function ZakenTable({ zaken, isLoaded, error, checkboxes, setCheckboxes }) {
+function ZakenTable({ zaken, isLoaded, error, checkboxes, setCheckboxes, canUpdateZaak=false }) {
     const [selectAll, setSelectAll] = useState(false);
 
     if (error) {
@@ -56,35 +48,20 @@ function ZakenTable({ zaken, isLoaded, error, checkboxes, setCheckboxes }) {
                 </tr>
             </thead>
             <tbody>
-                {zaken.map(zaak => (
-                <tr
-                  key={zaak.url}
-                  className={"zaak-record" + (!zaak.available ? " zaak-record--disabled" : "")}
-                >
-                    <td>
-                      <CheckboxInput
-                          checked={checkboxes[zaak.url] || false}
-                          name={zaak.url}
-                          onChange={(e) => {
-                              const isChecked = !checkboxes[zaak.url];
-                              setCheckboxes({...checkboxes, [zaak.url]: isChecked});
-                              if (!isChecked) {
-                                  setSelectAll(false);
-                              }
-                          }}
-                          disabled={!zaak.available}
-                      />
-                    </td>
-                    <td>{ zaak.identificatie }</td>
-                    <td>{displayZaaktype(zaak.zaaktype)}</td>
-                    <td>{zaak.omschrijving}</td>
-                    <td>{ zaak.looptijd }</td>
-                    <td>{ zaak.verantwoordelijkeOrganisatie }</td>
-                    <td>{ zaak.resultaat ? zaak.resultaat.resultaattype.omschrijving : '-' }</td>
-                    <td>{ zaak.resultaat ? zaak.resultaat.resultaattype.archiefactietermijn : '-'}</td>
-                    <td>{ zaak.zaaktype.processtype ? zaak.zaaktype.processtype.nummer : "-" }</td>
-                    <td>{ zaak.relevanteAndereZaken.length > 0 ? "Ja" : "Nee" }</td>
-                </tr>
+                {zaken.map((zaak, key) => (
+                    <ZaakRecord
+                        key={key}
+                        zaak={zaak}
+                        isChecked={checkboxes[zaak.url] || false}
+                        onCheckboxUpdate={(e) => {
+                            const isChecked = !checkboxes[zaak.url];
+                            setCheckboxes({...checkboxes, [zaak.url]: isChecked});
+                            if (!isChecked) {
+                              setSelectAll(false);
+                            }
+                        }}
+                        canUpdate={canUpdateZaak}
+                    />
                 ))}
 
             </tbody>
