+++ src/archiefvernietigingscomponent/sass/components/_zaak-record.scss
@@ -13,4 +13,8 @@
     background-color: $color-background;
     opacity: 0.65;
   }
+
+  .zaak-record__edit {
+    cursor: pointer;
+  }
 }
